refactoring of Context: moved simplifying functions from LwjglContext to Context
